refactor(workflow): extract rollover storage helpers
